PAS 2024 Rework : CAN & User Config Update (#741)
* Initialize CAN-OD new parameters

* Added dynamic parameters to the pedal assist struct, initialized in vc config.

* Can vehicule interface completed

* Added dynamic parameters to PAS source files

* User Config Updated

* CAN and UserConfig Completed

+++ src/GNR2_Firmware/revA/grn2_ra6t2_ftexboard/FTEX/inc/application_vehicle/vc_parameters_e_cells.h
@@ -68,9 +68,9 @@
 #define RUNTIME_PAS_SPEED_THRESHOLD                 5.0f    // Speed threshold to enable startup PAS power in km/h 
 #define STARTUP_PAS_SPEED_THRESHOLD                 1.0f    // Speed threshold to enable runtime PAS power in km/h
 #define TORQUE_STARTUP_VALUE_THRESHOLD    (uint16_t)1      // Torque value (%) that needs to be provided to have a startup detection
-#define STARTUP_PULSE_NUMBER              (uint32_t)1       // Number of pulses that needs to be detected to trigger startup detection
+#define STARTUP_PULSE_NUMBER              (uint16_t)1       // Number of pulses that needs to be detected to trigger startup detection
 #define STARTUP_TIME_WINDOW               (uint16_t)1000    // Time window (ms) in which the startup pulse number is counted
-#define RUNTIME_PULSE_NUMBER              (uint32_t)1       // Number of pulses that needs to be detected to trigger runtime detection
+#define RUNTIME_PULSE_NUMBER              (uint16_t)1       // Number of pulses that needs to be detected to trigger runtime detection
 #define RUNTIME_TIME_WINDOW               (uint16_t)400     // Time window (ms) in which the runtime pulse number is counted
 
 // Flag used to detect PAS with cadence AND torque.
@@ -100,8 +100,8 @@
 // false : option will be decativated/bypassed
 #define TORQUE_SCALING_PEDAL_RPM    false
 
-#define MIN_RPM_SCALING    (uint16_t)2          // Minimum pedaling RPM where the torque gain scaling starts
-#define MAX_RPM_SCALING    (uint16_t)50         // Maximum pedaling RPM where the torque gain scaling stops
+#define MIN_RPM_SCALING              2.0f       // Minimum pedaling RPM where the torque gain scaling starts
+#define MAX_RPM_SCALING              50.0f      // Maximum pedaling RPM where the torque gain scaling stops
 #define GAIN_AT_MIN_RPM              100.0f     // Torque scaling gain (in %) applied at min RPM
 #define GAIN_AT_MAX_RPM              300.0f     // Torque scaling gain (in %) applied at max RPM
